LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-waits-for-activate.https-expected.txt


PASS Navigation fetch events should wait for the activate event to complete.
FAIL Subresource fetch events should wait for the activate event to complete. assert_equals: active worker should be in activated state expected "activated" but got "activating"
